[collector] enable python-less collector implementation.
